Re-implement in TypeScript, add monorepo support (#4)
This PR re-implements this Action in TypeScript and adds support for monorepos. I have copied `MetaMask/action-create-release-pr` to the greatest extent possible, both in terms of structure and shared utilities, now living in `@metamask/action-utils`.

The responsibilities of this action are to compute the release notes from the changelog(s) of the released package(s), and then to create a GitHub release. We accomplish this by using `@metamask/auto-changelog`, and using the monorepo vs. polyrepo logic from `action-create-release-pr`.