Add a way to set a custom NPM registry
This change makes it possible to test actions that involve publishing of
NPM packages because it allows us to point to a private NPM registry.
For instance, you could run the action by saying:

``` yaml
- uses: MetaMask/action-publish-release@v2
  with:
    npm-registry-uri-fragment: //npm.fury.io/yourusernamehere/
    npm-token: ${{ secrets.NPM_TOKEN }}
  env:
    GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
```

and now when a package is accessed or published, Gemfury will be used
instead of the official NPM registry.