chore: rename release strategy to version strategy
This is the first step required to introduce
an actual release strategy. The current
RELEASE_STRATEGY env var was actually just
defining if monorepo workspaces should
consider version independently (and a tag
for each version was then created). But
a release strategy should also enable
the creation of independent releases
with independent release notes.