feat: add monorepo release strategy
Besides versioning strategy, we also
want to support the ability to publish
independent releases for each workspace.
This commit allows that. All it takes is
add a releasingStrategy property to the
release.config.json. One thing to remember
is that release strategy is only relevant
when versioning strategy is set to independent.