Create independent releases per workspace (#66)
* chore: rename release strategy to version strategy

This is the first step required to introduce
an actual release strategy. The current
RELEASE_STRATEGY env var was actually just
defining if monorepo workspaces should
consider version independently (and a tag
for each version was then created). But
a release strategy should also enable
the creation of independent releases
with independent release notes.

* feat: add monorepo release strategy

Besides versioning strategy, we also
want to support the ability to publish
independent releases for each workspace.
This commit allows that. All it takes is
add a releasingStrategy property to the
release.config.json. One thing to remember
is that release strategy is only relevant
when versioning strategy is set to independent.