Update tests to freeze snaps repo to 0.32.2
The latest release of the snaps packages seems to have been issued from
a branch that is not the main branch. Hence, the test that involves the
snaps repo is currently failing. At some point we should rewrite this
test so that it is not based on a real-world repo, but for the moment,
let's freeze snaps to a tag where we know that this test works.