List package versions in releases for monorepos with independent versioning (#79)
Currently, the notes for an independently-versioned monorepo release
lists all of the packages that are included in that release, but does
not list the versions of those packages that are being published. This
makes reviewing past releases difficult as we have to cross-reference
package tags with monorepo tags to determine those versions.

This commit fixes that. For instance, instead of seeing the following
for a GitHub release:

> # 59.0.0
>
> ## @metamask/base-controller
>
> (release notes for base-controller)
>
> ## @metamask/controller-utils
>
> (release notes for controller-utils)

we should now see:

> # 59.0.0
>
> ## @metamask/base-controller 4.0.0
>
> (release notes for base-controller)
>
> ## @metamask/controller-utils 5.0.0
>
> (release notes for controller-utils)