fix: skip packages with version 0.0.0
When adding a new package to a monorepo, this script will fail unless
the new package is included in the next release. It sees the initial
version of `0.0.0` as "new" because it doesn't match the version on npm
(which does not exist).

The script has been updated to skip packages with the version `0.0.0`
to better accommodate this scenario.