fix: Don't attempt publishing non-latest published versions
When deciding whether to attempt publishing a package, this action
would previously compare the latest published version with the current
version. If they don't match, it would try to publish the package.
However, that version may have been published already, just not as the
latest version.

This is especially problematic for backports within monorepos. This
would result in errors if any package was already published but not at
the latest version.

The script has been updated to check whether the given package version
was published at all. It no longer matters whether it was latest or not.

This also fixes the "update-packages-test", which tests with an old
version of the snaps monorepo. It was broken because all of the
packages being tested have had releases since the last successful run.
The test has been updated to add debug logs, and the expected number of
packages published has been changed to "2".

The test bumps 3 packages and decrements one. One of the bumped
packages was the `examples` package, which is private, so that has
always been ignored. The expected number of packages has changed from 3
to 2 because the decremented package is now skipped (it was already
published).