Use workflow-created commit for release tag
When this action is run, a tag is created for the release (via `gh
release create`) that points to the latest commit on the `main` branch.
This has the potential to be incorrect because a new commit could show
up on `main` in the brief window of time between when a release PR is
merged and when this action starts running.

To solve this problem, we pass a `--target` option to `gh release
create` which points to the commit that kicked off the release workflow
(which we can assume is the release commit).